feat: userCreator 가 유저 타입도 받을 수 있도록 변경
