+++ groove-security/auth/src/main/java/org/bogus/groove/domain/user/UserRegisterParam.java
@@ -8,4 +8,5 @@ import lombok.RequiredArgsConstructor;
 public class UserRegisterParam {
     private final String email;
     private final String password;
+    private final UserType userType = UserType.GROOVE;
 }
